scripts/requirements.txt
openai
requests
python-dotenv==1.0.1